+++ nf_core/pipelines/lint/nextflow_config.py
@@ -80,7 +80,7 @@ def nextflow_config(self) -> Dict[str, List[str]]:
     * ``params.nf_required_version``: The old method for specifying the minimum Nextflow version. Replaced by ``manifest.nextflowVersion``
     * ``params.container``: The old method for specifying the dockerhub container address. Replaced by ``process.container``
     * ``igenomesIgnore``: Changed to ``igenomes_ignore``
-    * ``params.max_cpus``: Old method of specifying the maximum number of CPUs a process can request. Replaced by native Nextflow `resourceLimits`directive.
+    * ``params.max_cpus``: Old method of specifying the maximum number of CPUs a process can request. Replaced by native Nextflow `resourceLimits`directive in config files.
     * ``params.max_memory``: Old method of specifying the maximum number of memory can request. Replaced by native Nextflow `resourceLimits`directive.
     * ``params.max_time``: Old method of specifying the maximum number of CPUs can request. Replaced by native Nextflow `resourceLimits`directive.
 
@@ -149,7 +149,13 @@ def nextflow_config(self) -> Dict[str, List[str]]:
         ["params.input"],
     ]
     # Throw a warning if these are missing
-    config_warn = [["manifest.mainScript"], ["timeline.file"], ["trace.file"], ["report.file"], ["dag.file"]]
+    config_warn = [
+        ["manifest.mainScript"],
+        ["timeline.file"],
+        ["trace.file"],
+        ["report.file"],
+        ["dag.file"],
+    ]
     # Old depreciated vars - fail if present
     config_fail_ifdefined = [
         "params.nf_required_version",
@@ -207,7 +213,11 @@ def nextflow_config(self) -> Dict[str, List[str]]:
             )
 
     # Remove field that should be ignored according to the linting config
-    ignore_configs = self.lint_config.get("nextflow_config", []) if self.lint_config is not None else []
+    ignore_configs = (
+        self.lint_config.get("nextflow_config", [])
+        if self.lint_config is not None
+        else []
+    )
 
     for cfs in config_fail:
         for cf in cfs: