+++ tests/pipelines/test_launch.py
@@ -17,9 +17,13 @@ class TestLaunch(TestPipelines):
     def setUp(self) -> None:
         super().setUp()
         self.nf_params_fn = Path(self.pipeline_dir, "nf-params.json")
-        self.launcher = nf_core.pipelines.launch.Launch(self.pipeline_dir, params_out=self.nf_params_fn)
+        self.launcher = nf_core.pipelines.launch.Launch(
+            self.pipeline_dir, params_out=self.nf_params_fn
+        )
 
-    @mock.patch.object(nf_core.pipelines.launch.Launch, "prompt_web_gui", side_effect=[True])
+    @mock.patch.object(
+        nf_core.pipelines.launch.Launch, "prompt_web_gui", side_effect=[True]
+    )
     @mock.patch.object(nf_core.pipelines.launch.Launch, "launch_web_gui")
     def test_launch_pipeline(self, mock_webbrowser, mock_lauch_web_gui):
         """Test the main launch function"""
@@ -34,10 +38,14 @@ class TestLaunch(TestPipelines):
         # Try and to launch, return with error
         assert self.launcher.launch_pipeline() is False
 
-    @mock.patch.object(nf_core.pipelines.launch.Launch, "prompt_web_gui", side_effect=[True])
+    @mock.patch.object(
+        nf_core.pipelines.launch.Launch, "prompt_web_gui", side_effect=[True]
+    )
     @mock.patch.object(nf_core.pipelines.launch.Launch, "launch_web_gui")
     @mock.patch.object(nf_core.pipelines.launch.Confirm, "ask", side_effect=[False])
-    def test_launch_file_exists_overwrite(self, mock_webbrowser, mock_lauch_web_gui, mock_confirm):
+    def test_launch_file_exists_overwrite(
+        self, mock_webbrowser, mock_lauch_web_gui, mock_confirm
+    ):
         """Test that we detect an existing params file and we overwrite it"""
         # Make an empty params file to be overwritten
         open(self.nf_params_fn, "a").close()
@@ -47,7 +55,14 @@ class TestLaunch(TestPipelines):
     def test_get_pipeline_schema(self):
         """Test loading the params schema from a pipeline"""
         self.launcher.get_pipeline_schema()
-        assert len(self.launcher.schema_obj.schema["$defs"]["input_output_options"]["properties"]) > 2
+        assert (
+            len(
+                self.launcher.schema_obj.schema["$defs"]["input_output_options"][
+                    "properties"
+                ]
+            )
+            > 2
+        )
 
     @with_temporary_folder
     def test_make_pipeline_schema(self, tmp_path):
@@ -58,10 +73,21 @@ class TestLaunch(TestPipelines):
         )
         create_obj.init_pipeline()
         Path(test_pipeline_dir, "nextflow_schema.json").unlink()
-        self.launcher = nf_core.pipelines.launch.Launch(test_pipeline_dir, params_out=self.nf_params_fn)
+        self.launcher = nf_core.pipelines.launch.Launch(
+            test_pipeline_dir, params_out=self.nf_params_fn
+        )
         self.launcher.get_pipeline_schema()
-        assert len(self.launcher.schema_obj.schema["$defs"]["input_output_options"]["properties"]) >= 2
-        assert self.launcher.schema_obj.schema["$defs"]["input_output_options"]["properties"]["outdir"] == {
+        assert (
+            len(
+                self.launcher.schema_obj.schema["$defs"]["input_output_options"][
+                    "properties"
+                ]
+            )
+            >= 2
+        )
+        assert self.launcher.schema_obj.schema["$defs"]["input_output_options"][
+            "properties"
+        ]["outdir"] == {
             "type": "string",
             "format": "directory-path",
             "description": "The output directory where the results will be saved. You have to use absolute paths to storage on Cloud infrastructure.",
@@ -91,8 +117,13 @@ class TestLaunch(TestPipelines):
         self.launcher.get_pipeline_schema()
         self.launcher.set_schema_inputs()
         self.launcher.merge_nxf_flag_schema()
-        assert self.launcher.schema_obj.schema["allOf"][0] == {"$ref": "#/$defs/coreNextflow"}
-        assert "-resume" in self.launcher.schema_obj.schema["$defs"]["coreNextflow"]["properties"]
+        assert self.launcher.schema_obj.schema["allOf"][0] == {
+            "$ref": "#/$defs/coreNextflow"
+        }
+        assert (
+            "-resume"
+            in self.launcher.schema_obj.schema["$defs"]["coreNextflow"]["properties"]
+        )
 
     def test_ob_to_questionary_string(self):
         """Check converting a python dict to a pyenquirer format - simple strings"""
@@ -101,14 +132,21 @@ class TestLaunch(TestPipelines):
             "default": "data/*{1,2}.fastq.gz",
         }
         result = self.launcher.single_param_to_questionary("input", sc_obj)
-        assert result == {"type": "input", "name": "input", "message": "", "default": "data/*{1,2}.fastq.gz"}
+        assert result == {
+            "type": "input",
+            "name": "input",
+            "message": "",
+            "default": "data/*{1,2}.fastq.gz",
+        }
 
     @mock.patch("questionary.unsafe_prompt", side_effect=[{"use_web_gui": "Web based"}])
     def test_prompt_web_gui_true(self, mock_prompt):
         """Check the prompt to launch the web schema or use the cli"""
         assert self.launcher.prompt_web_gui() is True
 
-    @mock.patch("questionary.unsafe_prompt", side_effect=[{"use_web_gui": "Command line"}])
+    @mock.patch(
+        "questionary.unsafe_prompt", side_effect=[{"use_web_gui": "Command line"}]
+    )
     def test_prompt_web_gui_false(self, mock_prompt):
         """Check the prompt to launch the web schema or use the cli"""
         assert self.launcher.prompt_web_gui() is False
@@ -123,17 +161,23 @@ class TestLaunch(TestPipelines):
         assert exc_info.value.args[0].startswith("Web launch response not recognised:")
 
     @mock.patch(
-        "nf_core.utils.poll_nfcore_web_api", side_effect=[{"api_url": "foo", "web_url": "bar", "status": "recieved"}]
+        "nf_core.utils.poll_nfcore_web_api",
+        side_effect=[{"api_url": "foo", "web_url": "bar", "status": "recieved"}],
     )
     @mock.patch("webbrowser.open")
     @mock.patch("nf_core.utils.wait_cli_function")
-    def test_launch_web_gui(self, mock_poll_nfcore_web_api, mock_webbrowser, mock_wait_cli_function):
+    def test_launch_web_gui(
+        self, mock_poll_nfcore_web_api, mock_webbrowser, mock_wait_cli_function
+    ):
         """Check the code that opens the web browser"""
         self.launcher.get_pipeline_schema()
         self.launcher.merge_nxf_flag_schema()
         assert self.launcher.launch_web_gui() is None
 
-    @mock.patch("nf_core.utils.poll_nfcore_web_api", side_effect=[{"status": "error", "message": "foo"}])
+    @mock.patch(
+        "nf_core.utils.poll_nfcore_web_api",
+        side_effect=[{"status": "error", "message": "foo"}],
+    )
     def test_get_web_launch_response_error(self, mock_poll_nfcore_web_api):
         """Test polling the website for a launch response - status error"""
         with pytest.raises(AssertionError) as exc_info:
@@ -145,14 +189,22 @@ class TestLaunch(TestPipelines):
         """Test polling the website for a launch response - status error"""
         with pytest.raises(AssertionError) as exc_info:
             self.launcher.get_web_launch_response()
-        assert exc_info.value.args[0].startswith("Web launch GUI returned unexpected status (foo): ")
+        assert exc_info.value.args[0].startswith(
+            "Web launch GUI returned unexpected status (foo): "
+        )
 
-    @mock.patch("nf_core.utils.poll_nfcore_web_api", side_effect=[{"status": "waiting_for_user"}])
+    @mock.patch(
+        "nf_core.utils.poll_nfcore_web_api",
+        side_effect=[{"status": "waiting_for_user"}],
+    )
     def test_get_web_launch_response_waiting(self, mock_poll_nfcore_web_api):
         """Test polling the website for a launch response - status waiting_for_user"""
         assert self.launcher.get_web_launch_response() is False
 
-    @mock.patch("nf_core.utils.poll_nfcore_web_api", side_effect=[{"status": "launch_params_complete"}])
+    @mock.patch(
+        "nf_core.utils.poll_nfcore_web_api",
+        side_effect=[{"status": "launch_params_complete"}],
+    )
     def test_get_web_launch_response_missing_keys(self, mock_poll_nfcore_web_api):
         """Test polling the website for a launch response - complete, but missing keys"""
         with pytest.raises(AssertionError) as exc_info:
@@ -175,7 +227,9 @@ class TestLaunch(TestPipelines):
         ],
     )
     @mock.patch.object(nf_core.pipelines.launch.Launch, "sanitise_web_response")
-    def test_get_web_launch_response_valid(self, mock_poll_nfcore_web_api, mock_sanitise):
+    def test_get_web_launch_response_valid(
+        self, mock_poll_nfcore_web_api, mock_sanitise
+    ):
         """Test polling the website for a launch response - complete, valid response"""
         self.launcher.get_pipeline_schema()
         assert self.launcher.get_web_launch_response() is True
@@ -185,11 +239,9 @@ class TestLaunch(TestPipelines):
         self.launcher.get_pipeline_schema()
         self.launcher.nxf_flags["-name"] = ""
         self.launcher.schema_obj.input_params["igenomes_ignore"] = "true"
-        self.launcher.schema_obj.input_params["max_cpus"] = "12"
         self.launcher.sanitise_web_response()
         assert "-name" not in self.launcher.nxf_flags
         assert self.launcher.schema_obj.input_params["igenomes_ignore"] is True
-        assert self.launcher.schema_obj.input_params["max_cpus"] == 12
 
     def test_ob_to_questionary_bool(self):
         """Check converting a python dict to a pyenquirer format - booleans"""
@@ -262,7 +314,10 @@ class TestLaunch(TestPipelines):
 
     def test_ob_to_questionary_pattern(self):
         """Check converting a python dict to a questionary format - with pattern"""
-        sc_obj = {"type": "string", "pattern": "^([a-zA-Z0-9_\\-\\.]+)@([a-zA-Z0-9_\\-\\.]+)\\.([a-zA-Z]{2,5})$"}
+        sc_obj = {
+            "type": "string",
+            "pattern": "^([a-zA-Z0-9_\\-\\.]+)@([a-zA-Z0-9_\\-\\.]+)\\.([a-zA-Z]{2,5})$",
+        }
         result = self.launcher.single_param_to_questionary("email", sc_obj)
         assert result["type"] == "input"
         assert result["validate"]("test@email.com") is True
@@ -295,7 +350,10 @@ class TestLaunch(TestPipelines):
         self.launcher.nxf_flags["-name"] = "Test_Workflow"
         self.launcher.nxf_flags["-resume"] = True
         self.launcher.build_command()
-        assert self.launcher.nextflow_cmd == f'nextflow run {self.pipeline_dir} -name "Test_Workflow" -resume'
+        assert (
+            self.launcher.nextflow_cmd
+            == f'nextflow run {self.pipeline_dir} -name "Test_Workflow" -resume'
+        )
 
     def test_build_command_params(self):
         """Test the functionality to build a nextflow command - params supplied"""
@@ -303,13 +361,18 @@ class TestLaunch(TestPipelines):
         self.launcher.schema_obj.input_params.update({"input": "custom_input"})
         self.launcher.build_command()
         # Check command
-        assert self.launcher.nextflow_cmd == f'nextflow run {self.pipeline_dir} -params-file "{self.nf_params_fn}"'
+        assert (
+            self.launcher.nextflow_cmd
+            == f'nextflow run {self.pipeline_dir} -params-file "{self.nf_params_fn}"'
+        )
         # Check saved parameters file
         with open(self.nf_params_fn) as fh:
             try:
                 saved_json = json.load(fh)
             except json.JSONDecodeError as e:
-                raise UserWarning(f"Unable to load JSON file '{self.nf_params_fn}' due to error {e}")
+                raise UserWarning(
+                    f"Unable to load JSON file '{self.nf_params_fn}' due to error {e}"
+                )
 
         assert saved_json == {"input": "custom_input"}
 
@@ -319,4 +382,7 @@ class TestLaunch(TestPipelines):
         self.launcher.get_pipeline_schema()
         self.launcher.schema_obj.input_params.update({"input": "custom_input"})
         self.launcher.build_command()
-        assert self.launcher.nextflow_cmd == f'nextflow run {self.pipeline_dir} --input "custom_input"'
+        assert (
+            self.launcher.nextflow_cmd
+            == f'nextflow run {self.pipeline_dir} --input "custom_input"'
+        )